abstract json blockly example
