+++ lib/blockly/examples/json_blockly/storage_key.ts
@@ -1,0 +1,1 @@
+export const storageKey = "jsonGeneratorWorkspace";
