--- lib/typescript_generator/mod.ts
@@ -1,3 +1,0 @@
-export * from "./typescript_generator.ts";
-export * from "./storage_key.ts";
-export * from "./toolbox.ts";
