--- lib/typescript_generator/storage_key.ts
@@ -1,1 +1,0 @@
-export const storageKey = "typescriptGeneratorWorkspace";
