--- lib/typescript_generator/toolbox.ts
@@ -1,33 +1,0 @@
-export const toolbox = {
-  "kind": "flyoutToolbox",
-  "contents": [
-    {
-      "kind": "block",
-      "type": "object",
-    },
-    {
-      "kind": "block",
-      "type": "member",
-    },
-    {
-      "kind": "block",
-      "type": "math_number",
-    },
-    {
-      "kind": "block",
-      "type": "text",
-    },
-    {
-      "kind": "block",
-      "type": "logic_boolean",
-    },
-    {
-      "kind": "block",
-      "type": "logic_null",
-    },
-    {
-      "kind": "block",
-      "type": "lists_create_with",
-    },
-  ],
-};
