+++ routes/index.tsx
@@ -1,9 +1,9 @@
-import TypeScriptBlocklyIsland from "#/islands/typescript_blockly_island.tsx";
+import BlocklyIsland from "../islands/blockly_island.tsx";
 
 export default function Home() {
   return (
     <main>
-      <TypeScriptBlocklyIsland />
+      <BlocklyIsland />
     </main>
   );
 }
